pyvuebot/templates/task_manager/requirements.txt
fastapi
python-telegram-bot
pydantic
uvicorn
supabase
python-dotenv